feat: add navigation & Header template
