+++ src/presentation/components/Template/Header/styles.scss
@@ -1,0 +1,21 @@
+.Container {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 1rem;
+    background-color: #333;
+  
+    .logo {
+      font-size: 1.5rem;
+      color: white;
+      width: 50px;
+      height: 50px;
+  
+      &:hover {
+        color: lightgray;
+        cursor: pointer;
+      }
+    }
+  
+  }
+
