+++ src/presentation/components/index.tsx
@@ -1,4 +1,5 @@
 export { default as Button } from './Button'
 export { default as Input } from './Input'
 export { default as FormLoaderStatus } from './FormLoaderStatus'
-export { default as Navigation } from './Template/Header/Navigation'
+export { default as Navigation } from './Template/Navigation'
+export { default as Header } from './Template/Header'
